# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ListingModel": {
        "properties": {
          "disposition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disposition"
          },
          "disposition_raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disposition Raw Text"
          },
          "favourited_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favourited At"
          },
          "first_seen_at": {
            "title": "First Seen At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location_raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Raw Text"
          },
          "price": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "price_drop_from": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Drop From"
          },
          "resolved_location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "sibling_sources": {
            "items": {
              "$ref": "#/components/schemas/SiblingSourceModel"
            },
            "title": "Sibling Sources",
            "type": "array"
          },
          "size_m2": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size M2"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "viewed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewed At"
          }
        },
        "required": [
          "id",
          "source",
          "url",
          "title",
          "location_raw_text",
          "resolved_location",
          "size_m2",
          "disposition_raw_text",
          "disposition",
          "first_seen_at",
          "viewed_at",
          "favourited_at",
          "price",
          "price_drop_from",
          "sibling_sources"
        ],
        "title": "ListingModel",
        "type": "object"
      },
      "ListingsArrivedEvent": {
        "properties": {
          "new": {
            "title": "New",
            "type": "integer"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "profile_id",
          "new"
        ],
        "title": "ListingsArrivedEvent",
        "type": "object"
      },
      "LocationModel": {
        "properties": {
          "cast_obce": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "cislo_orientacni": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cislo Orientacni"
          },
          "cislo_popisne": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cislo Popisne"
          },
          "kraj": {
            "$ref": "#/components/schemas/PlaceModel"
          },
          "mestska_cast": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "obec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "obvod": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "okres": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "ulice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "kraj",
          "okres",
          "obec",
          "obvod",
          "mestska_cast",
          "cast_obce",
          "ulice",
          "cislo_popisne",
          "cislo_orientacni"
        ],
        "title": "LocationModel",
        "type": "object"
      },
      "PlaceModel": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "PlaceModel",
        "type": "object"
      },
      "PortalHealthModel": {
        "properties": {
          "checked_at": {
            "title": "Checked At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "listing_count": {
            "title": "Listing Count",
            "type": "integer"
          },
          "portal": {
            "title": "Portal",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "broken",
              "zero_results"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "portal",
          "status",
          "error",
          "listing_count",
          "checked_at"
        ],
        "title": "PortalHealthModel",
        "type": "object"
      },
      "ProfileModel": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "enabled"
        ],
        "title": "ProfileModel",
        "type": "object"
      },
      "RunCountsModel": {
        "properties": {
          "disappeared": {
            "title": "Disappeared",
            "type": "integer"
          },
          "new": {
            "title": "New",
            "type": "integer"
          },
          "price_drops": {
            "title": "Price Drops",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "new",
          "price_drops",
          "disappeared"
        ],
        "title": "RunCountsModel",
        "type": "object"
      },
      "RunFinishedEvent": {
        "properties": {
          "counts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunCountsModel"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Error"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "partial",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "profile_id",
          "status"
        ],
        "title": "RunFinishedEvent",
        "type": "object"
      },
      "RunProgressEvent": {
        "properties": {
          "listing_count": {
            "title": "Listing Count",
            "type": "integer"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "scraper": {
            "title": "Scraper",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "broken",
              "zero_results"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "profile_id",
          "scraper",
          "status",
          "listing_count"
        ],
        "title": "RunProgressEvent",
        "type": "object"
      },
      "RunStartedEvent": {
        "properties": {
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "profile_id"
        ],
        "title": "RunStartedEvent",
        "type": "object"
      },
      "RunTriggeredModel": {
        "properties": {
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "profile_id"
        ],
        "title": "RunTriggeredModel",
        "type": "object"
      },
      "SiblingSourceModel": {
        "properties": {
          "source": {
            "title": "Source",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "source",
          "url"
        ],
        "title": "SiblingSourceModel",
        "type": "object"
      },
      "TriggerRunBody": {
        "properties": {
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          }
        },
        "required": [
          "profile_id"
        ],
        "title": "TriggerRunBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "rentczecher sidecar API",
    "version": "0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/events": {
      "get": {
        "operationId": "stream_events_v1_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events"
      }
    },
    "/v1/health": {
      "get": {
        "operationId": "health_v1_health_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortalHealthModel"
                  },
                  "title": "Response Health V1 Health Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Health"
      }
    },
    "/v1/profiles": {
      "get": {
        "operationId": "list_profiles_v1_profiles_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileModel"
                  },
                  "title": "Response List Profiles V1 Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Profiles"
      }
    },
    "/v1/profiles/{profile_id}/listings": {
      "get": {
        "operationId": "list_listings_v1_profiles__profile_id__listings_get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "default": "new",
              "enum": [
                "new",
                "all"
              ],
              "title": "Filter",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ListingModel"
                  },
                  "title": "Response List Listings V1 Profiles  Profile Id  Listings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Listings"
      }
    },
    "/v1/profiles/{profile_id}/listings/{listing_id}/viewed": {
      "patch": {
        "operationId": "mark_viewed_v1_profiles__profile_id__listings__listing_id__viewed_patch",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "title": "Listing Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Viewed"
      }
    },
    "/v1/runs": {
      "post": {
        "operationId": "trigger_run_v1_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerRunBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunTriggeredModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Run"
      }
    }
  }
}